// src/features/bloggers-platform/domain/dto/post-domain.dto.ts
export class CreatePostDomainDto {
  title: string;
  shortDescription: string;
  content: string;
  blogId: string;
  createdAt?: string;
  blogName: string;
}

export class UpdatePostDomainDto {
  title: string;
  shortDescription: string;
  content: string;
  blogId: string;
  blogName: string;
}